update add trans that reflect amount to the balance field in user table

+++ app/src/main/java/objects/User.java
@@ -3,8 +3,7 @@ package objects;
 public class User {
     public String email;
     public String userID;
-
-    public int balance;
+    public double balance;
 
     public User(){
